finalized everything

+++ src/main/java/com/codecool/dealership/service/CarService.java
@@ -1,34 +1,43 @@
 package com.codecool.dealership.service;
+import com.codecool.dealership.dto.CarDto;
 import com.codecool.dealership.model.Car;
 import com.codecool.dealership.repository.CarRepository;
-import com.codecool.dealership.repository.CustomerRepository;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CarService {
 
     private CarRepository carRepository;
 
-    private CustomerRepository customerRepository;
-
     public CarService(CarRepository carRepository) {
         this.carRepository = carRepository;
     }
 
-    public List<Car> retrieveCars() {
-        return carRepository.findAll();
+    public CarDto getCarById(Long id) {
+        Car car = carRepository.findById(id).orElse(null);
+        if (car == null) {
+            return null;
+        }
+        return new CarDto(car);
     }
 
-    public Car createCar(Car car) {
-        return carRepository.save(car);
+    public List<CarDto> retrieveCars() {
+        return carRepository.findAll().stream().map(CarDto::new).collect(Collectors.toList());
     }
 
-    public void deleteCar(Long id) {
-        carRepository.deleteById(id);
+    public CarDto createAndUpdateCar(Car car) {
+        return new CarDto(carRepository.save(car));
     }
 
-    public Car updateCar(Car car) {
-        return carRepository.save(car);
+    public String deleteCar(Long id) {
+        if (carRepository.existsById(id)) {
+           carRepository.deleteById(id);
+           return "Car successfully deleted";
+        } else {
+            return "Car not found";
+        }
     }
+
 }
